Changes to items of "sign-up-pitch" box/container
Text of "hero-main" is bolded with changed color to make page more appealing.

Color of text of "hero-scnd" is changed to make page more appealing.

Space is allocated around each item of "hero" box.

Color and design of sign-up button is changed to it more appealing.

Space is allocated around each item of "sign-up-pitch" box.