Changes to alignment of items in "Main-content"
Heading of "Main-conent" is centered & bolded to increase user engagement.

Items of Main-content is centered to increase user retention.

Space is allocated between each item of Main-content to enhance user experience.

Dimensions of images in Main-content are fixed.

Borders of images are modified to increase user retention.

Textual content of perks are aligned beneath images to improve user engagement.